Refactor: rename test utils, add coverage to gitignore
