+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
 dist/
+coverage/
 .idea/
 .DS_Store
 .vscode
