+++ packages/internal/dex/sdk/src/lib/poolUtils/fetchValidPools.test.ts
@@ -12,7 +12,7 @@ import {
   TEST_RPC_URL,
   TEST_V3_CORE_FACTORY_ADDRESS,
   WETH_TEST_TOKEN,
-} from '../../utils/testUtils';
+} from '../../test/utils';
 
 jest.mock('@ethersproject/contracts');
