+++ packages/internal/dex/sdk/src/lib/poolUtils/generatePossiblePoolsFromERC20Pairs.test.ts
@@ -7,7 +7,7 @@ import {
   USDC_TEST_TOKEN,
   WETH_TEST_TOKEN,
   uniqBy,
-} from '../../utils/testUtils';
+} from '../../test/utils';
 
 describe('generatePoolsFromTokenPairs', () => {
   describe('when given one TokenPair and one CommonRoutingTokens', () => {
