+++ packages/internal/dex/sdk/src/test/utils.test.ts
@@ -1,4 +1,4 @@
-import { uniqBy } from './testUtils';
+import { uniqBy } from './utils';
 
 describe('uniqBy', () => {
   describe('when given an array of numbers with Math.floor func', () => {
